requirements.txt
pywin32
Pillow
numpy
pyautogui
rapidocr-onnxruntime
